Splitted ui and logic in two projects
