--- Blazor/Data/CharacterClass.cs
@@ -1,11 +1,0 @@
-namespace GloomhavenAbilityManager.Blazor.Data
-{
-    public class CharacterClass
-    {
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-
-        public int NumberOfCards { get; set; }
-    }
-}
