--- Blazor/Data/IAbilityCardService.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace GloomhavenAbilityManager.Blazor.Data
-{
-    public interface IAbilityCardService
-    {
-        Task<AbilityCardInfo> GetCardAsync(int id);
-
-        Task<IEnumerable<AbilityCardInfo>> GetCardsAsync(int characterClassId);
-    }
-}
