--- Blazor/Data/ICharacterService.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace GloomhavenAbilityManager.Blazor.Data
-{
-    public interface ICharacterService
-    {
-        Character GetCharacter(int id);
-        Task<IEnumerable<Character>> GetCharactersAsync();
-    }
-}
