+++ Blazor/Pages/Card.razor
@@ -1,6 +1,7 @@
 @page "/card/{CardIdParameter}"
 
-@using GloomhavenAbilityManager.Blazor.Data
+@using GloomhavenAbilityManager.Logic.Contracts
+@using GloomhavenAbilityManager.Logic.Data
 @inject IAbilityCardService CardService
 @inject ICharacterClassService ClassService
