+++ Blazor/Pages/Char.razor
@@ -1,6 +1,7 @@
 @page "/char/{CharacterIdParameter}"
 
-@using GloomhavenAbilityManager.Blazor.Data
+@using GloomhavenAbilityManager.Logic.Contracts
+@using GloomhavenAbilityManager.Logic.Data
 @inject IAbilityCardService CardService
 @inject ICharacterClassService ClassService
 @inject ICharacterService CharacterService
