+++ Blazor/Pages/SelectableCard.razor
@@ -1,4 +1,5 @@
-@using GloomhavenAbilityManager.Blazor.Data
+@using GloomhavenAbilityManager.Logic.Contracts
+@using GloomhavenAbilityManager.Logic.Data
 @inject IAbilityCardService CardService
 @inject ICharacterClassService ClassService
