+++ Blazor/Shared/NavMenu.razor
@@ -1,4 +1,5 @@
-﻿@using GloomhavenAbilityManager.Blazor.Data
+﻿@using GloomhavenAbilityManager.Logic.Contracts
+@using GloomhavenAbilityManager.Logic.Data
 @inject ICharacterService CharacterService
 
 <div class="top-row pl-4 navbar navbar-dark">
