+++ Logic/Contracts/IAbilityCardRepository.cs
@@ -1,7 +1,8 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using GloomhavenAbilityManager.Logic.Data;
 
-namespace GloomhavenAbilityManager.Blazor.Data
+namespace GloomhavenAbilityManager.Logic.Contracts
 {
     public interface IAbilityCardRepository
     {
