+++ Logic/Contracts/ICharacterService.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using GloomhavenAbilityManager.Logic.Data;
+
+namespace GloomhavenAbilityManager.Logic.Contracts
+{
+    public interface ICharacterService
+    {
+        Character GetCharacter(int id);
+        Task<IEnumerable<Character>> GetCharactersAsync();
+    }
+}
