+++ Logic/Data/CharacterClass.cs
@@ -1,0 +1,11 @@
+namespace GloomhavenAbilityManager.Logic.Data
+{
+    public class CharacterClass
+    {
+        public int Id { get; set; }
+
+        public string Name { get; set; }
+
+        public int NumberOfCards { get; set; }
+    }
+}
